Feedback, readme, restructure files and directories
